replace tenant email class
